refactor(tmux): move presence tests into a sibling module
Keep the production module below the inline-test size invariant after adding
boundary-audit coverage.
